+++ src/main/groovy/ru/vyarus/gradle/plugin/quality/task/InitQualityConfigTask.groovy
@@ -5,6 +5,7 @@ import org.gradle.api.DefaultTask
 import org.gradle.api.tasks.Input
 import org.gradle.api.tasks.TaskAction
 import ru.vyarus.gradle.plugin.quality.ConfigLoader
+import ru.vyarus.gradle.plugin.quality.util.SpotbugsUtils
 
 /**
  * Task copies default configs to user directory (quality.configDir) for customization.
@@ -22,12 +23,12 @@ class InitQualityConfigTask extends DefaultTask {
     boolean override
 
     InitQualityConfigTask() {
-        group  = 'build setup'
+        group = 'build setup'
         description = 'Copies default quality plugin configuration files for customization'
     }
 
     @TaskAction
     void run() {
-        new ConfigLoader(project).initUserConfigs(override)
+        new ConfigLoader(project).initUserConfigs(override, SpotbugsUtils.isPluginEnabled(project))
     }
 }