+++ src/test/groovy/ru/vyarus/gradle/plugin/quality/ExcludeKitTest.groovy
@@ -44,7 +44,7 @@ class ExcludeKitTest extends AbstractKitTest {
         result.task(":check").outcome == TaskOutcome.SUCCESS
         def output = result.output
         output.contains "2 Checkstyle rule violations were found in 1 files"
-        output.contains "1 (0 / 1 / 0) FindBugs violations were found in 1 files"
+        output.contains "1 (0 / 1 / 0) SpotBugs violations were found in 1 files"
         output.contains "7 PMD rule violations were found in 1 files"
         output.contains "14 (0 / 5 / 9) CodeNarc violations were found in 1 files"
     }