+++ src/test/groovy/ru/vyarus/gradle/plugin/quality/QualityPluginLegacyKitTest.groovy
@@ -42,7 +42,7 @@ class QualityPluginLegacyKitTest extends AbstractKitTest {
         result.task(":check").outcome == TaskOutcome.SUCCESS
         result.output.contains('CodeNarc rule violations were found')
         result.output.contains('Checkstyle rule violations were found')
-        result.output.contains('FindBugs rule violations were found')
+        result.output.contains('SpotBugs rule violations were found')
         result.output.contains('PMD rule violations were found')
     }
 }