+++ src/test/groovy/ru/vyarus/gradle/plugin/quality/SpotbugsPluginsKitTest.groovy
@@ -5,16 +5,16 @@ import org.gradle.testkit.runner.TaskOutcome
 
 /**
  * @author Vyacheslav Rusakov
- * @since 21.12.2015
+ * @since 20.02.2018
  */
-class FindbugsPluginsKitTest extends AbstractKitTest {
+class SpotbugsPluginsKitTest extends AbstractKitTest {
 
-    def "Check findbugs plugins"() {
+    def "Check spotbugs plugins"() {
         setup:
         build("""
             plugins {
                 id 'java'
-                id 'findbugs'
+                id 'com.github.spotbugs'
                 id 'ru.vyarus.quality'
             }
 
@@ -22,12 +22,11 @@ class FindbugsPluginsKitTest extends AbstractKitTest {
                 checkstyle false
                 pmd false
                 strict false
-                spotbugs false
             }
 
             repositories { mavenCentral() }
             dependencies {
-                findbugsPlugins 'com.mebigfatguy.fb-contrib:fb-contrib:6.6.0'
+                spotbugsPlugins 'com.mebigfatguy.fb-contrib:fb-contrib:6.6.0'
             }
         """)
 
@@ -39,10 +38,10 @@ class FindbugsPluginsKitTest extends AbstractKitTest {
 
         then: "all plugins detect violations"
         result.task(":check").outcome == TaskOutcome.SUCCESS
-        result.output.contains('FindBugs rule violations were found')
+        result.output.contains('SpotBugs rule violations were found')
     }
 
-    def "Check findbugs plugins 2"() {
+    def "Check spotbugs plugins 2"() {
         setup:
         build("""
             plugins {
@@ -54,13 +53,12 @@ class FindbugsPluginsKitTest extends AbstractKitTest {
                 checkstyle false
                 pmd false
                 strict false
-                spotbugs false
             }
 
             repositories { mavenCentral() }
             afterEvaluate {
                 dependencies {
-                    findbugsPlugins 'com.mebigfatguy.fb-contrib:fb-contrib:6.6.0'
+                    spotbugsPlugins 'com.mebigfatguy.fb-contrib:fb-contrib:6.6.0'
                 }
             }
         """)
@@ -73,6 +71,6 @@ class FindbugsPluginsKitTest extends AbstractKitTest {
 
         then: "all plugins detect violations"
         result.task(":check").outcome == TaskOutcome.SUCCESS
-        result.output.contains('FindBugs rule violations were found')
+        result.output.contains('SpotBugs rule violations were found')
     }
 }