+++ src/test/groovy/ru/vyarus/gradle/plugin/quality/task/InitQualityConfigTaskKitTest.groovy
@@ -16,6 +16,34 @@ class InitQualityConfigTaskKitTest extends AbstractKitTest {
             plugins {
                 id 'java'
                 id 'ru.vyarus.quality'
+                id 'com.github.spotbugs'
+            }
+
+            quality {
+                configDir = 'config'
+            }
+        """)
+
+        when: "run init configs task"
+        BuildResult result = run('initQualityConfig')
+
+        then: "configs copied"
+        result.task(':initQualityConfig').outcome == TaskOutcome.SUCCESS
+        file('config/checkstyle/checkstyle.xml').exists()
+        file('config/codenarc/codenarc.xml').exists()
+        file('config/spotbugs/exclude.xml').exists()
+        file('config/spotbugs/html-report-style.xsl').exists()
+        !file('config/findbugs').exists()
+        file('config/pmd/pmd.xml').exists()
+    }
+
+    def "Check configs findbugs init"() {
+        setup:
+        build("""
+            plugins {
+                id 'java'
+                id 'ru.vyarus.quality'
+                //spotbugs not registered!
             }
 
             quality {
@@ -32,6 +60,7 @@ class InitQualityConfigTaskKitTest extends AbstractKitTest {
         file('config/codenarc/codenarc.xml').exists()
         file('config/findbugs/exclude.xml').exists()
         file('config/findbugs/html-report-style.xsl').exists()
+        !file('config/spotbugs').exists()
         file('config/pmd/pmd.xml').exists()
     }
 }